Improve bool'ness (and add explicit static_cast<int>() for printf like operations to be compatible of a bool being a C++ object) (#6325)
git-svn-id: https://svn.osgeo.org/gdal/trunk@33093 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/apps/gdalsrsinfo.cpp
@@ -384,7 +384,8 @@ CPLErr PrintSRS( const OGRSpatialReference &oSRS,
         return CE_None;
 
     CPLDebug( "gdalsrsinfo", "PrintSRS( oSRS, %s, %d, %d )\n",
-              pszOutputType, bPretty, bPrintSep );
+              pszOutputType, static_cast<int>(bPretty),
+              static_cast<int>(bPrintSep) );
 
     char *pszOutput = NULL;
